backend/requirements.txt
asgiref==3.11.1
django==5.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg==3.3.2
psycopg-binary==3.3.2
python-dotenv==1.2.1
sqlparse==0.5.5
typing-extensions==4.15.0
tzdata==2025.3
gunicorn==25.1.0
